// SaintMichael/src/app/models/curso.ts
export interface Curso {
    nombreInstitucion: string;
    grupoCurso: string;
    nombreCurso : string;
    fondo: string;
    idProfesor: any;
    estadoCurso: boolean;
}

export interface Curso2 {
    idCurso: any;
    nombreInstitucion: string;
    grupoCurso: string;
    nombreCurso : string;
    fondo: string;
    idProfesor: any;
    estadoCurso: boolean;
}